#1097 - scope tab, remove unused component
