+++ src/frontend/src/components/CheckList.tsx
@@ -3,12 +3,11 @@
  * See the LICENSE file in the repository root folder for details.
  */
 
-import PageBlock from '../layouts/PageBlock';
 import Checkbox from '@mui/material/Checkbox';
 import FormControl from '@mui/material/FormControl';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import Typography from '@mui/material/Typography';
-import { ReactNode, useState } from 'react';
+import { useState } from 'react';
 import { useCheckDescriptionBullet } from '../hooks/description-bullets.hooks';
 import { useAuth } from '../hooks/auth.hooks';
 import { Tooltip } from '@mui/material';
@@ -25,12 +24,11 @@ export type CheckListItem = {
 
 interface CheckListProps {
   title: string;
-  headerRight?: ReactNode;
   items: CheckListItem[];
   isDisabled: boolean;
 }
 
-const CheckList: React.FC<CheckListProps> = ({ title, headerRight, items, isDisabled }) => {
+const CheckList: React.FC<CheckListProps> = ({ title, items, isDisabled }) => {
   const auth = useAuth();
   const { isLoading, mutateAsync } = useCheckDescriptionBullet();
   const [showConfirm, setShowConfirm] = useState<boolean>(false);
@@ -54,7 +52,10 @@ const CheckList: React.FC<CheckListProps> = ({ title, headerRight, items, isDisa
   });
 
   return (
-    <PageBlock title={title} headerRight={headerRight}>
+    <>
+      <Typography variant="h5" sx={{ mb: 1 }}>
+        {title}
+      </Typography>
       <FormControl>
         {items.map((check, idx) => (
           <FormControlLabel
@@ -106,7 +107,7 @@ const CheckList: React.FC<CheckListProps> = ({ title, headerRight, items, isDisa
       >
         <Typography>Are you sure you want to mark this completed task as NOT completed?</Typography>
       </NERModal>
-    </PageBlock>
+    </>
   );
 };
